function1/requirements.txt
azure-servicebus
azure-cosmos
azure-identity
prettytable